+++ src/BlazorGPT/Pages/MemoriesPage.razor.css
